Separate file and add mutiple save and read
